userfeaturegenerator input schema fix

+++ src/caaswx/spark/_transformers/userfeaturegenerator.py
@@ -10,7 +10,7 @@ from pyspark.sql.functions import window
 from pyspark.sql.types import (
     TimestampType,
     StringType,
-    LongType,
+    IntegerType,
 )
 from pyspark.sql.window import Window
 
@@ -335,7 +335,7 @@ class UserFeatureGenerator(SparkNativeTransformer):
 
     sch_dict = {
         "SM_TIMESTAMP": ["SM_TIMESTAMP", TimestampType()],
-        "SM_EVENTID": ["SM_EVENTID", LongType()],
+        "SM_EVENTID": ["SM_EVENTID", IntegerType()],
         "SM_RESOURCE": ["SM_RESOURCE", StringType()],
         "SM_CLIENTIP": ["SM_CLIENTIP", StringType()],
         "SM_USERNAME": ["SM_USERNAME", StringType()],
